Add dark mode docs and prepare for release

+++ build/build.js
@@ -20,7 +20,9 @@ async function build() {
   log('Cleaning "dist/, docs/static/assets/paper.css" folder...');
 
   rimraf.sync('dist', { disableGlob: true });
-  fs.unlinkSync(paperDocsPath);
+  if (fs.existsSync(paperDocsPath)) {
+    fs.unlinkSync(paperDocsPath);
+  }
 
   log('Compiling SCSS to CSS, entrypoint:', entrypoint);
